Macro to run CGI on mac / linux

+++ test/runTests.py
@@ -166,8 +166,7 @@ def GET_TESTS(testNum = 0):
 
     # ------- GET : 200 (with CGI)
     index += 1
-    # if (platform == "darwin" and (testNum == 0 or index == int(testNum))):
-    if (testNum == 0 or index == int(testNum)):
+    if (platform == "darwin" and (testNum == 0 or index == int(testNum))):
         r = requests.get("http://localhost:7777/index.bla")
         assertResponse(r, 200, index)
     index += 1
@@ -294,7 +293,7 @@ def POST_TESTS(testNum = 0):
     
     # ------- POST CHUNKED - 200/201 - NO CGI
     index += 1
-    if (testNum == 0 or index == int(testNum)):
+    if (platform == "darwin" and (testNum == 0 or index == int(testNum))):
         if os.path.exists("www/newFile"): os.remove("www/newFile")
         payload = "14\r\nabcdefghijklmnopqrst\r\nA\r\n0123456789\r\n0\r\n\r\n"
         hd = {'Transfer-Encoding': 'chunked'}
@@ -380,6 +379,7 @@ def DELETE_TESTS(testNum = 0):
 # -----------------------------------------------------------------------------
 
 
+print("\n       Platform = " + platform)
 
 if (len(sys.argv) == 1):
     GET_TESTS()
